Split out projects into its own route
