+++ src/Projects/img/scmla-conference-sessions.png
